aws-agentcore-runtime/requirements.txt
bedrock-agentcore
strands-agents
pydantic
pydantic_core